requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.50.0
streamlit>=1.0.0
transformers>=4.0.0
gTTS>=2.2.0
pygame>=2.0.0
requests>=2.25.0
scipy>=1.5.0
pytest>=6.0.0